requirements.txt
anthropic
requests
beautifulsoup4
streamlit
python-dotenv